Add support for building on Linux
This commit adds support for building on Linux, currently there are
some ifdefs and a TODO in the code.

TODO: Currently there is a need to include OVR_CAPI from the src-directory of the oculus sdk. I don't know of a good way to handle this

I think it would be nice to get rid of the ifdef in the headerfile
over the function taking HWND by either changing it to a void* or
moving it over to the cpp-file.