Use ovrlib release version if no debug present
Oculus SDK ships with only release version of the library pre-built.
They are possible to build from source, but this change makes it
possible to run the CMake configuration without the debug library.
Instead it will use the release version in the debug build as well.
This is bad form and the CMake script will emitt a warning urging the
user to build the debug version of ovrlib from source.

This closes #65